Implement prove RPC method
