+++ crates/sovereign-sdk/rollup-interface/src/state_machine/stf.rs
@@ -174,7 +174,8 @@ pub trait StateTransitionFunction<Vm: Zkvm, Da: DaSpec> {
         + AsRef<[u8]>
         + Debug
         + Sync
-        + Send;
+        + Send
+        + 'static;
 
     /// The initial params of the rollup.
     type GenesisParams;
@@ -199,7 +200,8 @@ pub trait StateTransitionFunction<Vm: Zkvm, Da: DaSpec> {
         + Serialize
         + DeserializeOwned
         + Send
-        + Sync;
+        + Sync
+        + 'static;
 
     /// The validity condition that must be verified outside of the Vm
     type Condition: ValidityCondition;
